entity changes

+++ library/src/main/java/com/library/usermanagement/repositories/UserDetailsRepository.java
@@ -1,11 +1,9 @@
 package com.library.usermanagement.repositories;
 
-import java.math.BigDecimal;
-
 import com.library.usermanagement.models.UserDetails;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDetailsRepository extends JpaRepository<UserDetails, BigDecimal>{
+public interface UserDetailsRepository extends JpaRepository<UserDetails, Long>{
     
 }
